+++ Source/Core/SceneLib/ImageBasedLightResource.cpp
@@ -56,7 +56,7 @@ namespace Selas
                                       filepath);
 
         void* fileData = nullptr;
-        uint32 fileSize = 0;
+        uint64 fileSize = 0;
         ReturnError_(File::ReadWholeFile(filepath.Ascii(), &fileData, &fileSize));
 
         AttachToBinary(resource->data, (uint8*)fileData, fileSize);
